requirements.txt
Flask
google-generativeai
python-dotenv
pdfplumber
flask-cors
vaderSentiment
gunicorn
